Move the application client to a shared location
