+++ tests/application_client/boilerplate_currency_utils.py
