+++ tests/standalone/test_name_version.py
@@ -3,7 +3,7 @@ from ragger.backend.interface import BackendInterface
 from application_client.boilerplate_command_sender import BoilerplateCommandSender
 from application_client.boilerplate_response_unpacker import unpack_get_app_and_version_response
 
-from utils import verify_version, verify_name
+from .utils import verify_version, verify_name
 
 # Test a specific APDU asking BOLOS (and not the app) the name and version of the current app
 def test_get_app_and_version(backend: BackendInterface) -> None:
