+++ tests/standalone/test_sign_cmd.py
@@ -7,7 +7,7 @@ from ragger.navigator.navigation_scenario import NavigateWithScenario
 from application_client.boilerplate_transaction import Transaction
 from application_client.boilerplate_command_sender import BoilerplateCommandSender, Errors
 from application_client.boilerplate_response_unpacker import unpack_get_public_key_response, unpack_sign_tx_response
-from utils import check_signature_validity
+from .utils import check_signature_validity
 
 # In this tests we check the behavior of the device when asked to sign a transaction
