+++ tests/standalone/test_version_cmd.py
@@ -3,7 +3,7 @@ from ragger.backend.interface import BackendInterface
 from application_client.boilerplate_command_sender import BoilerplateCommandSender
 from application_client.boilerplate_response_unpacker import unpack_get_version_response
 
-from utils import verify_version
+from .utils import verify_version
 
 # In this test we check the behavior of the device when asked to provide the app version
 def test_version(backend: BackendInterface) -> None:
